collector: WiFi-node cfg contract, flash headroom, host tests
WiFi node (docs/NODE_CONFIG.md §3, after cce02cf):
- an empty "readings" array is accepted: it marks the node alive
  (RemoteIngest::touch moves last-seen, not the values' age) and the cfg
  exchange still runs
- ingest body cap 6 KB, one shared accumulateBody() for /api/ingest and
  the /api/nodes/* POSTs
- a key with no config file answers nulls from GET /api/nodes/config,
  never 404; the handover sort moved into NodeCfgStore
- reported is filled from desired once a node applies with no report

Flash: sscanf linked newlib's vfscanf (~8.7 KB) for two dotted-quad
parses; src/utils/Ipv4Parse.h replaces it (stricter: 300 is refused,
not wrapped to 44). Store/API refactors bring the all-features C3 image
back under app0: 1,506,448 of 1,507,328 bytes.

Tests: test_nodecfg_collector.cpp (NodeCfgRules.h, 185 checks),
test_ipv4_parse.cpp, RemoteIngest::touch. NodeCfgRules.h: cast in
statusAfterApplied for -Wextra. BH1750 comment points at
node_common/NodeSensors.cpp. ARCHITECTURE §6 and NODE_CONFIG §3/§5/§7
updated.

+++ src/nodes/NodeCfgRules.h
@@ -70,7 +70,7 @@ static inline uint16_t nextRev(uint16_t rev) {
 /// rejected — the rejected rev is still the desired one, and the node is
 /// still (correctly) running its previous config.
 static inline uint8_t statusAfterApplied(uint8_t status, uint16_t rev, uint16_t applied) {
-    return applied >= rev ? ST_APPLIED : status;
+    return applied >= rev ? (uint8_t)ST_APPLIED : status;
 }
 
 /// Is there a config the node should be handed? Pending and behind. A
